refactor: 파일 이름 변경
- 스터디 불러오기 파일 이름 변경
- 배경화면에 따른 text className 파일 분리
