+++ src/components/StudyHomeComponents/backgrounds/RecentBackground.js
@@ -1,14 +1,10 @@
-import styles from '../StudyDataFetch.module.css';
+import styles from '../RecentDataFetch.module.css';
 import backgroundGreen from '../../../img/background/background_1.png';
 import backgroundYe from '../../../img/background/background_2.png';
 import backgroundBlu from '../../../img/background/background_3.png';
 import backgroundPink from '../../../img/background/background_4.png';
-// import backgroundTable from '../../img/background/background_5.png';
-// import backgroundSun from '../../img/background/background_6.png';
-// import backgroundRain from '../../img/background/background_7.png';
-// import backgroundPlan from '../../img/background/background_8.png';
 
-const styleMapping = {
+const RecentBackground = {
   [backgroundGreen]: {
     nameColor: styles.nameGreen,
     studyNameColor: styles.studyNameBlack,
@@ -46,4 +42,4 @@ const styleMapping = {
   },
 };
 
-export default styleMapping;
+export default RecentBackground;
